# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Vesalius-Dr Backend API description.",
        "title": "Swagger Vesalius-Dr Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/mobile_central_dr-1.0.0",
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DoctorAppUser"
                        }
                    }
                }
            }
        },
        "/user/add-machine-id": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "description": "AddMachineId Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PostMachineInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "description": "ChangePassword Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PostChangePasswordDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/email/{email}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DoctorAppUser"
                        }
                    }
                }
            }
        },
        "/user/first-time-change-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "description": "ChangePassword Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PostChangePasswordDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/userId/{userId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.DoctorAppUser"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.PostChangePasswordDto": {
            "type": "object",
            "required": [
                "newPassword",
                "oldPassword"
            ],
            "properties": {
                "newPassword": {
                    "type": "string",
                    "minLength": 1
                },
                "oldPassword": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "dto.PostMachineInfo": {
            "type": "object",
            "required": [
                "machineId"
            ],
            "properties": {
                "machineId": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "model.DoctorAppUser": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "contact_number": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "doctor_app_user_id": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "firstTimeLogin": {
                    "type": "boolean"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "machine_id": {
                    "type": "string"
                },
                "mcr": {
                    "type": "string"
                },
                "middle_name": {
                    "type": "string"
                },
                "nationality": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}